% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{google_language_codes}
\alias{google_language_codes}
\title{List Valid Google Language Codes}
\usage{
google_language_codes(x = c("latin", "latin-ext", "sinhala", "greek",
  "hebrew", "vietnamese", "cyrillic", "cyrillic-ext", "devanagari", "arabic",
  "khmer", "tamil", "greek-ext", "thai", "bengali", "gujarati", "oriya",
  "malayalam", "gurmukhi", "kannada", "telugu", "myanmar"))
}
\arguments{
\item{x}{Vector of potential Google language codes}
}
\seealso{
\code{\link[=google_font]{google_font()}}
}
