% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lighten_darken_color}
\alias{lighten_darken_color}
\alias{lighten_color}
\alias{darken_color}
\title{Generate lighter or darker version of a color}
\usage{
lighten_color(color_hex, strength = 0.7)

darken_color(color_hex, strength = 0.8)
}
\arguments{
\item{color_hex}{A character string representing a hex color}

\item{strength}{The "strength" of the blend with white or black,
0 low to 1 high.}
}
\description{
Produces a linear blend of the color with white or black.
}
