% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono.R
\name{mono_accent_inverse}
\alias{mono_accent_inverse}
\title{Monotone Accent Inverse Theme}
\usage{
mono_accent_inverse(base_color = "#cbf7ed", white_color = "#FFF",
  black_color = darken_color(base_color, 0.9), text_color = white_color,
  header_color = base_color, background_color = black_color,
  link_color = base_color, text_bold_color = base_color,
  text_slide_number_color = base_color, code_highlight_color = "#ffff88",
  code_inline_color = base_color, code_inline_background_color = NA,
  inverse_background_color = base_color, inverse_text_color = black_color,
  inverse_text_shadow = FALSE, inverse_header_color = black_color,
  title_slide_text_color = inverse_text_color,
  title_slide_background_color = inverse_background_color,
  title_slide_background_image = NA,
  left_column_subtle_color = darken_color(base_color, 0.5),
  left_column_selected_color = base_color,
  blockquote_left_color = darken_color(base_color, 0.5),
  table_border_color = "#666", table_row_border_color = "#ddd",
  table_row_even_background_color = darken_color(base_color, 0.7),
  text_font_google = NULL, text_font_family = "'Droid Serif'",
  text_font_weight = "normal",
  text_font_url = "https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic",
  text_font_family_fallback = "'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'",
  text_font_base = "serif", header_font_google = NULL,
  header_font_family = "'Yanone Kaffeesatz'", header_font_weight = "normal",
  header_font_url = "https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz",
  code_font_google = NULL, code_font_family = "'Source Code Pro'",
  code_font_url = "https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700",
  code_font_family_fallback = "'Lucida Console', Monaco",
  outfile = "xaringan-themed.css")
}
\arguments{
\item{base_color}{Monotone Base Color}

\item{white_color}{Color for "white". Monotone Light defaults to a very light
version of the \code{base_color}}

\item{black_color}{Color for "black". Monotone Light defaults to a dark,
nearly black version of the \code{base_color}}

\item{text_color}{Text Color}

\item{header_color}{Header Color}

\item{background_color}{Slide Background Color}

\item{link_color}{Link Color}

\item{text_bold_color}{Bold Text Color}

\item{text_slide_number_color}{Slide Number Color}

\item{code_highlight_color}{Code Line Highlight}

\item{code_inline_color}{Inline Code Color}

\item{code_inline_background_color}{Inline Code Background Color}

\item{inverse_background_color}{Inverse Background Color}

\item{inverse_text_color}{Inverse Text Color}

\item{inverse_text_shadow}{Enables Shadow on text of inverse slides}

\item{inverse_header_color}{Inverse Header Color}

\item{title_slide_text_color}{Title Slide Text Color}

\item{title_slide_background_color}{Title Slide Background Color}

\item{title_slide_background_image}{Title Slide Background Image URL}

\item{left_column_subtle_color}{Left Column Text (not last)}

\item{left_column_selected_color}{Left Column Current Selection}

\item{blockquote_left_color}{Blockquote Left Border Color}

\item{table_border_color}{Table top/bottom border}

\item{table_row_border_color}{Table row inner bottom border}

\item{table_row_even_background_color}{Table Even Row Background Color}

\item{text_font_google}{Use \code{google_font()} to specify body font}

\item{text_font_family}{Body Text Font Family}

\item{text_font_weight}{Body Text Font Weight}

\item{text_font_url}{Body Text Font URL(s)}

\item{text_font_family_fallback}{Body Text Font Fallbacks}

\item{text_font_base}{Body Text Base Font (Total Failure Fallback)}

\item{header_font_google}{Use \code{google_font()} to specify header font}

\item{header_font_family}{Header Font Family}

\item{header_font_weight}{Header Font Weight}

\item{header_font_url}{Header Font URL}

\item{code_font_google}{Use \code{google_font()} to specify code font}

\item{code_font_family}{Code Font Family}

\item{code_font_url}{Code Font URL}

\item{code_font_family_fallback}{Code Font Fallback}

\item{outfile}{Customized xaringan CSS output file name}
}
\description{
Monotone Accent Inverse Theme
}
