% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_xaringan}
\alias{scale_xaringan}
\alias{scale_xaringan_discrete}
\alias{scale_xaringan_fill_discrete}
\alias{scale_xaringan_color_discrete}
\alias{scale_xaringan_colour_discrete}
\alias{scale_xaringan_continuous}
\alias{scale_xaringan_fill_continuous}
\alias{scale_xaringan_color_continuous}
\alias{scale_xaringan_colour_continuous}
\title{Xaringan Themer ggplot2 Scales}
\usage{
scale_xaringan_discrete(aes_type = c("color", "colour", "fill"), ...,
  color = NULL, direction = 1, inverse = FALSE)

scale_xaringan_fill_discrete(..., color = NULL, direction = 1,
  inverse = FALSE)

scale_xaringan_color_discrete(..., color = NULL, direction = 1,
  inverse = FALSE)

scale_xaringan_colour_discrete(..., color = NULL, direction = 1,
  inverse = FALSE)

scale_xaringan_continuous(aes_type = c("color", "colour", "fill"), ...,
  color = NULL, begin = 0, end = 1, inverse = FALSE)

scale_xaringan_fill_continuous(..., color = NULL, begin = 0, end = 1,
  inverse = FALSE)

scale_xaringan_color_continuous(..., color = NULL, begin = 0,
  end = 1, inverse = FALSE)

scale_xaringan_colour_continuous(..., color = NULL, begin = 0,
  end = 1, inverse = FALSE)
}
\arguments{
\item{aes_type}{The type of aesthetic to which the scale is being applied.
One of "color", "colour", or "fill".}

\item{...}{Arguments passed on to the appropriate scale function, one of
\link[colorspace:scale_color_discrete_sequential]{colorspace::scale_color_discrete_sequential},
\link[colorspace:scale_color_continuous_sequential]{colorspace::scale_color_continuous_sequential},
\link[colorspace:scale_fill_discrete_sequential]{colorspace::scale_fill_discrete_sequential}, or
\link[colorspace:scale_fill_continuous_sequential]{colorspace::scale_fill_continuous_sequential}.}

\item{color}{A color value, in hex, to override the default color. Otherwise,
the primary color of the resulting scale is chosen from the xaringanthemer
slide styles.}

\item{direction}{Direction of the discrete scale. Use values less than 0 to
reverse the direction, e.g. \code{direction = -1}.}

\item{inverse}{If \code{color} is not supplied and \code{inverse = TRUE}, a primary
color is chosen to work well with the inverse slide styles, namely the
value of \code{inverse_header_color}}

\item{begin}{Number in the range of \code{[0, 1]} indicating to which point in the color scale the smallest data value should be mapped.}

\item{end}{Number in the range of \code{[0, 1]} indicating to which point in the color scale the largest data value should be mapped.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}} or
\code{\link[=ggplot2:discrete_scale]{ggplot2:discrete_scale()}}.}
}
\description{
\strong{Lifecycle:} \href{https://www.tidyverse.org/lifecycle/#experimental}{Experimental}.

Color and fill single-color scales for discrete and continuous values,
created using the primary accent color of the xaringanthemer styles.
}
