% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solarized_light.R
\name{solarized_light}
\alias{solarized_light}
\title{Solarized Light Theme}
\usage{
solarized_light(text_color = "#657b83", header_color = "#dc322f",
  background_color = "#fdf6e3", link_color = "#b58900",
  text_bold_color = "#d33682", text_slide_number_color = "#93a1a1",
  code_highlight_color = "#268bd240", code_inline_color = "#6c71c4",
  code_inline_background_color = NA, inverse_background_color = "#002b36",
  inverse_text_color = "#fdf6e3", inverse_text_shadow = FALSE,
  inverse_header_color = inverse_text_color,
  title_slide_text_color = inverse_text_color,
  title_slide_background_color = inverse_background_color,
  title_slide_background_image = NA, left_column_subtle_color = "#93a1a1",
  left_column_selected_color = "#586e75", blockquote_left_color = "#cb4b16",
  table_border_color = "#839496", table_row_border_color = "#839496",
  table_row_even_background_color = "#eee8d5", text_size = "20px",
  text_size_h1 = "55px", text_size_h2 = "45px", text_size_h3 = "35px",
  text_size_slide_number = "20px", text_size_code_inline = "20px",
  text_size_code = "18px", text_font_google = NULL,
  text_font_family = "'Droid Serif'", text_font_weight = "normal",
  text_font_url = "https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic",
  text_font_family_fallback = "'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'",
  text_font_base = "serif", header_font_google = NULL,
  header_font_family = "'Yanone Kaffeesatz'", header_font_weight = "normal",
  header_font_url = "https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz",
  code_font_google = NULL, code_font_family = "'Source Code Pro'",
  code_font_url = "https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700",
  code_font_family_fallback = "'Lucida Console', Monaco",
  outfile = "xaringan-themed.css")
}
\arguments{
\item{text_color}{Text Color, defaults to #657b83}

\item{header_color}{Header Color, defaults to #dc322f}

\item{background_color}{Slide Background Color, defaults to #fdf6e3}

\item{link_color}{Link Color, defaults to #b58900}

\item{text_bold_color}{Bold Text Color, defaults to #d33682}

\item{text_slide_number_color}{Slide Number Color, defaults to #93a1a1}

\item{code_highlight_color}{Code Line Highlight, defaults to #268bd240}

\item{code_inline_color}{Inline Code Color, defaults to #6c71c4}

\item{code_inline_background_color}{Inline Code Background Color, defaults to NA}

\item{inverse_background_color}{Inverse Background Color, defaults to #002b36}

\item{inverse_text_color}{Inverse Text Color, defaults to #fdf6e3}

\item{inverse_text_shadow}{Enables Shadow on text of inverse slides, defaults to \code{FALSE}}

\item{inverse_header_color}{Inverse Header Color, defaults to \code{inverse_text_color}}

\item{title_slide_text_color}{Title Slide Text Color, defaults to \code{inverse_text_color}}

\item{title_slide_background_color}{Title Slide Background Color, defaults to \code{inverse_background_color}}

\item{title_slide_background_image}{Title Slide Background Image URL, defaults to NA}

\item{left_column_subtle_color}{Left Column Text (not last), defaults to #93a1a1}

\item{left_column_selected_color}{Left Column Current Selection, defaults to #586e75}

\item{blockquote_left_color}{Blockquote Left Border Color, defaults to #cb4b16}

\item{table_border_color}{Table top/bottom border, defaults to #839496}

\item{table_row_border_color}{Table row inner bottom border, defaults to #839496}

\item{table_row_even_background_color}{Table Even Row Background Color, defaults to #eee8d5}

\item{text_size}{Slide Body Text Size, defaults to 20px}

\item{text_size_h1}{h1 Header Text Size, defaults to 55px}

\item{text_size_h2}{h2 Header Text Size, defaults to 45px}

\item{text_size_h3}{h3 Header Text Size, defaults to 35px}

\item{text_size_slide_number}{Slide Number Text Size, defaults to 20px}

\item{text_size_code_inline}{Inline Code Text Size, defaults to 20px}

\item{text_size_code}{Code Text Size, defaults to 18px}

\item{text_font_google}{Use \code{google_font()} to specify body font, defaults to \code{NULL}}

\item{text_font_family}{Body Text Font Family, defaults to 'Droid Serif'}

\item{text_font_weight}{Body Text Font Weight, defaults to normal}

\item{text_font_url}{Body Text Font URL(s), defaults to https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic}

\item{text_font_family_fallback}{Body Text Font Fallbacks, defaults to 'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'}

\item{text_font_base}{Body Text Base Font (Total Failure Fallback), defaults to serif}

\item{header_font_google}{Use \code{google_font()} to specify header font, defaults to \code{NULL}}

\item{header_font_family}{Header Font Family, defaults to 'Yanone Kaffeesatz'}

\item{header_font_weight}{Header Font Weight, defaults to normal}

\item{header_font_url}{Header Font URL, defaults to https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz}

\item{code_font_google}{Use \code{google_font()} to specify code font, defaults to \code{NULL}}

\item{code_font_family}{Code Font Family, defaults to 'Source Code Pro'}

\item{code_font_url}{Code Font URL, defaults to https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700}

\item{code_font_family_fallback}{Code Font Fallback, defaults to 'Lucida Console', Monaco}

\item{outfile}{Customized xaringan CSS output file name}
}
\description{
A light theme based around on the solarized color scheme.
  Works well with "\code{highlightStyle: solarized-dark}" or
  "\code{highlighStyle: solarized-light}".
}
\references{
\url{http://ethanschoonover.com/solarized}
}
\seealso{
Other Solarized themes: \code{\link{solarized_dark}}
}
