% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{theme_xaringan}
\alias{theme_xaringan}
\title{A Plot Theme for ggplot2 by xaringanthemer}
\usage{
theme_xaringan(
  text_color = NULL,
  background_color = NULL,
  accent_color = NULL,
  accent_secondary_color = NULL,
  ...
)
}
\arguments{
\item{text_color}{Color for text and foreground, inherits from \code{text_color}}

\item{background_color}{Color for background, inherits from
\code{background_color}}

\item{accent_color}{Color for titles and accents, inherits from
\code{header_color}}

\item{accent_secondary_color}{Color for secondary accents, inherits from
\code{text_bold_color}}

\item{...}{
  Arguments passed on to \code{\link[=theme_xaringan_base]{theme_xaringan_base}}
  \describe{
    \item{\code{set_ggplot_defaults}}{Should defaults be set for \pkg{ggplot2} \emph{geoms}?
Defaults to TRUE. To restore ggplot's defaults, or the previously set geom
defaults, see \code{\link[=theme_xaringan_restore_defaults]{theme_xaringan_restore_defaults()}}.}
    \item{\code{text_font}}{Font to use for text elements, passed to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}}, if available and \code{text_font_use_google} is
\code{TRUE}. Inherits from \code{text_font_family}.}
    \item{\code{text_font_use_google}}{Is \code{text_font} available on \href{https://fonts.google.com}{Google Fonts}?}
    \item{\code{text_font_size}}{Base text font size, inherits from \code{text_font_size}, or
defaults to 11.}
    \item{\code{title_font}}{Font to use for title elements, passed to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}}, if available and \code{title_font_use_google} is
\code{TRUE}. Inherits from \code{title_font_family}.}
    \item{\code{title_font_use_google}}{Is \code{title_font} available on \href{https://fonts.google.com}{Google Fonts}?}
    \item{\code{title_font_size}}{Base text font size, inherits from \code{title_font_size},
or defaults to 14.}
  }}
}
\value{
A ggplot2 theme
}
\description{
\strong{Lifecycle:} \href{https://www.tidyverse.org/lifecycle/#experimental}{Experimental}.

Creates \pkg{ggplot2} themes to match the xaringanthemer theme used in the
\pkg{xaringan} slides that seamlessly matches the "normal" slide colors and
styles.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  # Set xaringanthemer theme but save to tempfile
  style_duo_accent(outfile = tempfile())

  library(ggplot2)
  ggplot(iris) +
    aes(Petal.Length, Petal.Width) +
    geom_point() +
    theme_xaringan()
}
}
\seealso{
Other xaringanthemer ggplot2 themes: 
\code{\link{theme_xaringan_base}()},
\code{\link{theme_xaringan_inverse}()},
\code{\link{theme_xaringan_set_defaults}()}
}
\concept{xaringanthemer ggplot2 themes}
