% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{theme_xaringan_base}
\alias{theme_xaringan_base}
\title{The ggplot2 xaringanthemer base plot theme}
\usage{
theme_xaringan_base(text_color, background_color, ...,
  set_ggplot_defaults = TRUE, accent_color = NULL,
  accent_secondary_color = NULL, text_font = NULL,
  text_font_use_google = NULL, text_font_size = NULL,
  title_font = NULL, title_font_use_google = NULL,
  title_font_size = NULL)
}
\arguments{
\item{text_color}{Color for text and foreground}

\item{background_color}{Color for background}

\item{...}{Ignored}

\item{set_ggplot_defaults}{Should defaults be set for \link{ggplot2} \emph{geoms}?
Defaults to TRUE. To restore ggplot's defaults, or the previously set geom
defaults, see \code{\link[=theme_xaringan_restore_defaults]{theme_xaringan_restore_defaults()}}.}

\item{accent_color}{Color for titles and accents, inherits from
\code{header_color} or \code{text_color}. Used for the \code{title} base setting in
\code{\link[ggplot2:theme]{ggplot2::theme()}}, and additionally for setting the \code{color} or \code{fill} of
\link{ggplot2} geom defaults.}

\item{accent_secondary_color}{Color for secondary accents, inherits from
\code{text_bold_color} or \code{accent_color}. Used only when setting \link{ggplot2} geom
defaults.}

\item{text_font}{Font to use for text elements, passed to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}}, if available and \code{text_font_use_google} is
\code{TRUE}. Inherits from \code{text_font_family}.}

\item{text_font_use_google}{Is \code{text_font} available on \href{https://fonts.google.com}{Google Fonts}?}

\item{text_font_size}{Base text font size, inherits from \code{text_font_size}, or
defaults to 11.}

\item{title_font}{Font to use for title elements, passed to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}}, if available and \code{title_font_use_google} is
\code{TRUE}. Inherits from \code{title_font_family}.}

\item{title_font_use_google}{Is \code{title_font} available on \href{https://fonts.google.com}{Google Fonts}?}

\item{title_font_size}{Base text font size, inherits from \code{title_font_size},
or defaults to 14.}
}
\value{
A ggplot2 theme
}
\description{
Provides a base plot theme for \link{ggplot2} to match the \link{xaringan} slide theme
created by \link{xaringanthemer}. The theme is designed to create a general plot
style from two colors, a \code{background_color} and a \code{text_color} (or foreground
color). Also accepts an \code{accent_color} and an \code{accent_secondary_color} that are
\link{xaringanthemer} is not required for the base theme. Use
\code{\link[=theme_xaringan]{theme_xaringan()}} or \code{\link[=theme_xaringan_inverse]{theme_xaringan_inverse()}} in xaringan slides styled by
xaringanthemer for a plot theme that matches the slide style.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  library(ggplot2)
  ggplot(iris) +
    aes(Petal.Length, Petal.Width) +
    geom_point() +
    theme_xaringan_base(
      text_color = "#e1e5f2",
      background_color = "#021c25",
      accent_color = "#1f7a8c",
      set_ggplot_defaults = TRUE) +
      labs(title = "Basic Iris Plot",
           subtitle = "+ theme_xaringan_base()",
           caption = "{xaringanthemer}")

   ggplot(iris) +
    aes(Petal.Length, Petal.Width) +
    geom_point() +
    theme_xaringan_base(
      text_color = "#021c25",
      background_color = "#e1e5f2",
      accent_color = "#1f7a8c",
      set_ggplot_defaults = TRUE) +
      labs(title = "Basic Iris Plot",
           subtitle = "+ theme_xaringan_base()",
           caption = "{xaringanthemer}")
}

}
\seealso{
Other xaringanthemer ggplot2 themes: \code{\link{theme_xaringan_inverse}},
  \code{\link{theme_xaringan_set_defaults}},
  \code{\link{theme_xaringan}}
}
\concept{xaringanthemer ggplot2 themes}
