% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{theme_xaringan_inverse}
\alias{theme_xaringan_inverse}
\title{An Inverse Plot Theme for ggplot2 by xaringanthemer}
\usage{
theme_xaringan_inverse(text_color = NULL, background_color = NULL,
  accent_color = NULL, accent_secondary_color = NULL, ...)
}
\arguments{
\item{text_color}{Color for text and foreground, inherits from \code{text_color}}

\item{background_color}{Color for background, inherits from
\code{background_color}}

\item{accent_color}{Color for titles and accents, inherits from
\code{header_color}}

\item{accent_secondary_color}{Color for secondary accents, inherits from
\code{text_bold_color}}

\item{...}{Arguments passed on to \code{theme_xaringan_base}
\describe{
  \item{text_color}{Color for text and foreground}
  \item{background_color}{Color for background}
  \item{accent_color}{Color for titles and accents, inherits from
\code{header_color} or \code{text_color}. Used for the \code{title} base setting in
\code{\link[ggplot2:theme]{ggplot2::theme()}}, and additionally for setting the \code{color} or \code{fill} of
\link{ggplot2} geom defaults.}
  \item{accent_secondary_color}{Color for secondary accents, inherits from
\code{text_bold_color} or \code{accent_color}. Used only when setting \link{ggplot2} geom
defaults.}
  \item{set_ggplot_defaults}{Should defaults be set for \link{ggplot2} \emph{geoms}?
Defaults to TRUE. To restore ggplot's defaults, or the previously set geom
defaults, see \code{\link[=theme_xaringan_restore_defaults]{theme_xaringan_restore_defaults()}}.}
  \item{text_font}{Font to use for text elements, passed to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}}, if available and \code{text_font_use_google} is
\code{TRUE}. Inherits from \code{text_font_family}.}
  \item{text_font_use_google}{Is \code{text_font} available on \href{https://fonts.google.com}{Google Fonts}?}
  \item{text_font_size}{Base text font size, inherits from \code{text_font_size}, or
defaults to 11.}
  \item{title_font}{Font to use for title elements, passed to
\code{\link[sysfonts:font_add_google]{sysfonts::font_add_google()}}, if available and \code{title_font_use_google} is
\code{TRUE}. Inherits from \code{title_font_family}.}
  \item{title_font_use_google}{Is \code{title_font} available on \href{https://fonts.google.com}{Google Fonts}?}
  \item{title_font_size}{Base text font size, inherits from \code{title_font_size},
or defaults to 14.}
}}
}
\value{
A ggplot2 theme
}
\description{
A \link{ggplot2} xaringanthemer plot theme to seamlessly match the "inverse"
\link{xaringan} slide colors and styles as styled by \link{xaringanthemer}.
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  # Set xaringanthemer theme but save to tempfile
  duo_accent(outfile = tempfile())

  library(ggplot2)
  ggplot(iris) +
    aes(Petal.Length, Petal.Width) +
    geom_point() +
    theme_xaringan()
}

}
\seealso{
Other xaringanthemer ggplot2 themes: \code{\link{theme_xaringan_base}},
  \code{\link{theme_xaringan_set_defaults}},
  \code{\link{theme_xaringan}}
}
\concept{xaringanthemer ggplot2 themes}
