% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{theme_xaringan_set_defaults}
\alias{theme_xaringan_set_defaults}
\alias{xaringan_theme_restore_defaults}
\title{Set and Restore ggplot2 geom Defaults}
\usage{
theme_xaringan_set_defaults(text_color = NULL, background_color = NULL,
  accent_color = text_color, accent_secondary_color = accent_color,
  text_family = NULL)

xaringan_theme_restore_defaults()
}
\arguments{
\item{text_color}{Color for text and foreground, inherits from \code{text_color}}

\item{background_color}{Color for background, inherits from
\code{background_color}}

\item{accent_color}{Color for titles and accents, inherits from
\code{header_color}}

\item{accent_secondary_color}{Color for secondary accents, inherits from
\code{text_bold_color}}
}
\value{
Invisibly returns a list of the current ggplot2 geom defaults

Invisibly returns a list of the current ggplot2 geom defaults
}
\description{
Set \link{ggplot2} \emph{geom} defaults to match \code{\link[=theme_xaringan]{theme_xaringan()}} with
\code{theme_xaringan_set_defaults()} and restore the standard or previously-set
defaults with \code{theme_xaringan_restore_defaults()}. By default,
\code{theme_xaringan_set_defaults()} is run with \code{\link[=theme_xaringan]{theme_xaringan()}} or
\code{\link[=theme_xaringan_inverse]{theme_xaringan_inverse()}}.
}
\section{Functions}{
\itemize{
\item \code{xaringan_theme_restore_defaults}: Restore previous or standard \link{ggplot2} \emph{geom} defaults.
}}

\seealso{
Other xaringanthemer ggplot2 themes: \code{\link{theme_xaringan_base}},
  \code{\link{theme_xaringan_inverse}},
  \code{\link{theme_xaringan}}
}
\concept{xaringanthemer ggplot2 themes}
