% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_extra_css.R
\name{write_extra_css}
\alias{write_extra_css}
\title{Write Extra CSS arguments}
\usage{
write_extra_css(css, outfile = "xaringan-themer.css", append = TRUE)
}
\arguments{
\item{css}{A named list of CSS definitions each containing a named list
of CSS property-value pairs, i.e.
\code{list(".class-id" = list("css-property" = "value"))}}

\item{outfile}{Customized xaringan CSS output file name, default is "xaringan-themer.css"}

\item{append}{If \code{TRUE} output will be appended to \code{outfile}; otherwise,
it will overwrite the contents of \code{outfile}.}
}
\description{
Manually adds css elements to target \code{outfile}.
}
\section{css list}{

The \code{css} input must be a named list of css properties and values within a
named list of class identifiers, for example
\code{list(".class-id" = list("css-property" = "value"))}.
}

