% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringanthemer-deprecated.R
\name{xaringanthemer-deprecated}
\alias{xaringanthemer-deprecated}
\alias{write_xaringan_theme}
\alias{write_extra_css}
\alias{mono_light}
\alias{mono_dark}
\alias{mono_accent}
\alias{mono_accent_inverse}
\alias{duo}
\alias{duo_accent}
\alias{duo_accent_inverse}
\alias{solarized_light}
\alias{solarized_dark}
\title{Deprecated or renamed functions}
\usage{
write_xaringan_theme(...)

write_extra_css(...)

mono_light(...)

mono_dark(...)

mono_accent(...)

mono_accent_inverse(...)

duo(...)

duo_accent(...)

duo_accent_inverse(...)

solarized_light(...)

solarized_dark(...)
}
\arguments{
\item{...}{Argumets passed to the new or renamed functions.}
}
\value{
The result of the new or renamed function.
}
\description{
These functions in xaringanthemer have been deprecated or renamed.

Use \code{\link[=style_xaringan]{style_xaringan()}} instead of \code{write_xaringan_theme()}.

Use \code{\link[=style_extra_css]{style_extra_css()}} instead of \code{write_extra_css()}.

Use \code{\link[=style_mono_light]{style_mono_light()}} instead of \code{mono_light()}.

Use \code{\link[=style_mono_dark]{style_mono_dark()}} instead of \code{mono_dark()}.

Use \code{\link[=style_mono_accent]{style_mono_accent()}} instead of \code{mono_accent()}.

Use \code{\link[=style_mono_accent_inverse]{style_mono_accent_inverse()}} instead of \code{mono_accent_inverse()}.

Use \code{\link[=style_duo]{style_duo()}} instead of \code{duo()}.

Use \code{\link[=style_duo_accent]{style_duo_accent()}} instead of \code{duo_accent()}.

Use \code{\link[=style_duo_accent_inverse]{style_duo_accent_inverse()}} instead of \code{duo_accent_inverse()}.

Use \code{\link[=style_solarized_light]{style_solarized_light()}} instead of \code{solarized_light()}.

Use \code{\link[=style_solarized_dark]{style_solarized_dark()}} instead of \code{solarized_dark()}.
}
\keyword{internal}
