% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringanthemer-package.R
\docType{package}
\name{xaringanthemer-package}
\alias{xaringanthemer}
\alias{xaringanthemer-package}
\title{xaringanthemer: Xaringan CSS Theme Generator}
\description{
Create and edit your xaringan CSS files from within your R
R Markdown slide source. More technically, this package actually creates
valid remarkjs CSS themes that are easily used by xarangan.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/xaringanthemer}
  \item Report bugs at \url{https://github.com/gadenbuie/xaringanthemer/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{g.adenbuie@gmail.com}

}
\keyword{internal}
