% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{google_language_codes}
\alias{google_language_codes}
\title{List Valid Google Language Codes}
\usage{
google_language_codes(
  language_codes = c("latin", "latin-ext", "sinhala", "greek", "hebrew", "vietnamese",
    "cyrillic", "cyrillic-ext", "devanagari", "arabic", "khmer", "tamil", "greek-ext",
    "thai", "bengali", "gujarati", "oriya", "malayalam", "gurmukhi", "kannada", "telugu",
    "myanmar")
)
}
\arguments{
\item{language_codes}{Vector of potential Google language codes}
}
\value{
A vector of Google Font language codes matching \code{language_codes}.
}
\description{
Gives a list of valid Language Codes for Google Fonts, or
validates that the language codes given are valid.
}
\seealso{
\code{\link[=google_font]{google_font()}}
}
