% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_accent.R
\name{mono_accent}
\alias{mono_accent}
\title{Monotone Accent Theme}
\usage{
mono_accent(base_color = "#43418A", white_color = "#FFFFFF",
  black_color = "#272822", text_color = black_color,
  header_color = base_color, background_color = white_color,
  link_color = base_color, text_bold_color = base_color,
  text_slide_number_color = base_color,
  padding = "1rem 4rem 1rem 4rem", background_image = NULL,
  background_size = NULL, background_position = NULL,
  code_highlight_color = "rgba(255,255,0,0.5)",
  code_inline_color = base_color, code_inline_background_color = NULL,
  code_inline_font_size = "1em", inverse_background_color = base_color,
  inverse_text_color = white_color, inverse_text_shadow = FALSE,
  inverse_header_color = white_color,
  title_slide_text_color = inverse_text_color,
  title_slide_background_color = inverse_background_color,
  title_slide_background_image = NULL,
  title_slide_background_size = NULL,
  title_slide_background_position = NULL, footnote_color = NULL,
  footnote_font_size = "0.9em", footnote_position_bottom = "3em",
  left_column_subtle_color = apply_alpha(base_color, 0.6),
  left_column_selected_color = base_color,
  blockquote_left_border_color = apply_alpha(base_color, 0.5),
  table_border_color = "#666", table_row_border_color = "#ddd",
  table_row_even_background_color = lighten_color(base_color, 0.7),
  text_font_size = "20px", header_h1_font_size = "55px",
  header_h2_font_size = "45px", header_h3_font_size = "35px",
  header_background_auto = FALSE,
  header_background_color = header_color,
  header_background_text_color = background_color,
  header_background_padding = "2rem 4rem 1.5rem 4rem",
  header_background_content_padding_top = "7rem",
  header_background_ignore_classes = c("normal", "inverse", "title",
  "middle", "bottom"), text_slide_number_font_size = "0.9em",
  text_font_google = NULL, text_font_family = "'Droid Serif'",
  text_font_weight = "normal",
  text_font_url = "https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic",
  text_font_family_fallback = "'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'",
  text_font_base = "serif", header_font_google = NULL,
  header_font_family = "'Yanone Kaffeesatz'",
  header_font_weight = "normal",
  header_font_url = "https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz",
  code_font_google = NULL, code_font_family = "'Source Code Pro'",
  code_font_size = "0.9em",
  code_font_url = "https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700",
  code_font_family_fallback = "'Lucida Console', Monaco",
  extra_css = NULL, extra_fonts = NULL,
  outfile = "xaringan-themer.css")
}
\arguments{
\item{base_color}{Monotone Base Color, works best with a strong color.. Defaults to #43418A. Modifies multiple CSS classes or elements.}

\item{white_color}{Brightest color used. Defaults to #FFFFFF. Modifies multiple CSS classes or elements.}

\item{black_color}{Darkest color used. Defaults to #272822. Modifies multiple CSS classes or elements.}

\item{text_color}{Text Color. Defaults to \code{black_color}. Modifies the \code{body} element.}

\item{header_color}{Header Color. Defaults to \code{base_color}. Modifies the \code{h1, h2, h3} elements.}

\item{background_color}{Slide Background Color. Defaults to \code{white_color}. Modifies the \code{.remark-slide-content} class.}

\item{link_color}{Link Color. Defaults to \code{base_color}. Modifies the \code{a, a > code} elements.}

\item{text_bold_color}{Bold Text Color. Defaults to \code{base_color}. Modifies the \code{strong} element.}

\item{text_slide_number_color}{Slide Number Color. Defaults to \code{base_color}. Modifies the \code{.remark-slide-number} class.}

\item{padding}{Slide Padding in \code{top right [bottom left]} format. Defaults to 1rem 4rem 1rem 4rem. Modifies the \code{.remark-slide-content} class.}

\item{background_image}{Background image applied to each \emph{and every} slide. Set \code{title_slide_background_image = "none"} to remove the background image from the title slide.. Defaults to \code{NULL}. Modifies the \code{.remark-slide-content} class.}

\item{background_size}{Background image size, requires \code{background_image} to be set. If \code{background_image} is set, \code{background_size} will default to \code{cover} so the backround fills the screen. If both \code{background_image} and \code{background_position} are set, will default to 100 percent.. Defaults to \code{NULL}. Modifies the \code{.remark-slide-content} class.}

\item{background_position}{Background image position, requires \code{background_image} to be set, and it is recommended to adjust \code{background_size}.. Defaults to \code{NULL}. Modifies the \code{.remark-slide-content} class.}

\item{code_highlight_color}{Code Line Highlight. Defaults to rgba(255,255,0,0.5). Modifies the \code{.remark-code-line-highlighted} class.}

\item{code_inline_color}{Inline Code Color. Defaults to \code{base_color}. Modifies the \code{.remark-inline-code} class.}

\item{code_inline_background_color}{Inline Code Background Color. Defaults to \code{NULL}. Modifies the \code{.remark-inline-code} class.}

\item{code_inline_font_size}{Inline Code Text Font Size. Defaults to 1em. Modifies the \code{.remark-inline-code} class.}

\item{inverse_background_color}{Inverse Background Color. Defaults to \code{base_color}. Modifies the \code{.inverse} class.}

\item{inverse_text_color}{Inverse Text Color. Defaults to \code{white_color}. Modifies the \code{.inverse} class.}

\item{inverse_text_shadow}{Enables Shadow on text of inverse slides. Defaults to \code{FALSE}. Modifies the \code{.inverse} class.}

\item{inverse_header_color}{Inverse Header Color. Defaults to \code{white_color}. Modifies the \code{.inverse h1, .inverse h2, .inverse h3} classes.}

\item{title_slide_text_color}{Title Slide Text Color. Defaults to \code{inverse_text_color}. Modifies the \code{.title-slide} class.}

\item{title_slide_background_color}{Title Slide Background Color. Defaults to \code{inverse_background_color}. Modifies the \code{.title-slide} class.}

\item{title_slide_background_image}{Title Slide Background Image URL. Defaults to \code{NULL}. Modifies the \code{.title-slide} class.}

\item{title_slide_background_size}{Title Slide Background Image Size, defaults to "cover" if background image is set. Defaults to \code{NULL}. Modifies the \code{.title-slide} class.}

\item{title_slide_background_position}{Title Slide Background Image Position. Defaults to \code{NULL}. Modifies the \code{.title-slide} class.}

\item{footnote_color}{Footnote text color (if \code{NA}, then it will be the same color as \code{text_color}). Defaults to \code{NULL}. Modifies the \code{.footnote} class.}

\item{footnote_font_size}{Footnote font size. Defaults to 0.9em. Modifies the \code{.footnote} class.}

\item{footnote_position_bottom}{Footnote location from bottom of screen. Defaults to 3em. Modifies the \code{.footnote} class.}

\item{left_column_subtle_color}{Left Column Text (not last). Defaults to \code{apply_alpha(base_color, 0.6)}. Modifies the \code{.left-column h2, .left-column h3} classes.}

\item{left_column_selected_color}{Left Column Current Selection. Defaults to \code{base_color}. Modifies the \code{.left-column h2:last-of-type, .left-column h3:last-child} classes.}

\item{blockquote_left_border_color}{Blockquote Left Border Color. Defaults to \code{apply_alpha(base_color, 0.5)}. Modifies the \code{blockquote} element.}

\item{table_border_color}{Table top/bottom border. Defaults to #666. Modifies the \code{table: border-top, border-bottom} elements.}

\item{table_row_border_color}{Table row inner bottom border. Defaults to #ddd. Modifies the \code{table thead th: border-bottom} elements.}

\item{table_row_even_background_color}{Table Even Row Background Color. Defaults to \code{lighten_color(base_color, 0.7)}. Modifies the \code{thead, tfoot, tr:nth-child(even)} elements.}

\item{text_font_size}{Slide Body Text Font Size. Defaults to 20px. Modifies the \code{.remark-slide-content} class.}

\item{header_h1_font_size}{h1 Header Text Font Size. Defaults to 55px. Modifies the \code{.remark-slide-content h1} class.}

\item{header_h2_font_size}{h2 Header Text Font Size. Defaults to 45px. Modifies the \code{.remark-slide-content h2} class.}

\item{header_h3_font_size}{h3 Header Text Font Size. Defaults to 35px. Modifies the \code{.remark-slide-content h3} class.}

\item{header_background_auto}{Add background under slide title automatically for h1 header elements. If not enabled, use \code{class: header_background} to enable.. Defaults to \code{FALSE}.}

\item{header_background_color}{Background Color for h1 Header with Background. Defaults to \code{header_color}. Modifies the \code{.remark-slide-content h1} class.}

\item{header_background_text_color}{Text Color for h1 Header with Background. Defaults to \code{background_color}. Modifies the \code{.remark-slide-content h1} class.}

\item{header_background_padding}{Padding for h1 Header with Background. Defaults to 2rem 4rem 1.5rem 4rem. Modifies the \code{.remark-slide-content h1} class.}

\item{header_background_content_padding_top}{Top Padding for Content in Slide with Header with Background. Defaults to 7rem. Modifies the \code{.remark-slide-content} class.}

\item{header_background_ignore_classes}{Slide Classes Where Header with Background will not be Applied. Defaults to \code{c('normal', 'inverse', 'title', 'middle', 'bottom')}. Modifies the \code{.remark-slide-content} class.}

\item{text_slide_number_font_size}{Slide Number Text Font Size. Defaults to 0.9em. Modifies the \code{.remark-slide-number} class.}

\item{text_font_google}{Use \code{google_font()} to specify body font. Defaults to \code{NULL}. Modifies the \code{body} element.}

\item{text_font_family}{Body Text Font Family. Defaults to 'Droid Serif'. Modifies the \code{body} element.}

\item{text_font_weight}{Body Text Font Weight. Defaults to normal. Modifies the \code{body} element.}

\item{text_font_url}{Body Text Font URL(s). Defaults to https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic. Modifies the \code{@import url()} elements.}

\item{text_font_family_fallback}{Body Text Font Fallbacks. Defaults to 'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'. Modifies the \code{body} element.}

\item{text_font_base}{Body Text Base Font (Total Failure Fallback). Defaults to serif. Modifies the \code{body} element.}

\item{header_font_google}{Use \code{google_font()} to specify header font. Defaults to \code{NULL}. Modifies the \code{body} element.}

\item{header_font_family}{Header Font Family. Defaults to 'Yanone Kaffeesatz'. Modifies the \code{h1, h2, h3} elements.}

\item{header_font_weight}{Header Font Weight. Defaults to normal. Modifies the \code{h1, h2, h3} elements.}

\item{header_font_url}{Header Font URL. Defaults to https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz. Modifies the \code{@import url} elements.}

\item{code_font_google}{Use \code{google_font()} to specify code font. Defaults to \code{NULL}. Modifies the \code{body} element.}

\item{code_font_family}{Code Font Family. Defaults to 'Source Code Pro'. Modifies the \code{.remark-code, .remark-inline-code} classes.}

\item{code_font_size}{Code Text Font Size. Defaults to 0.9em. Modifies the \code{.remark-inline} class.}

\item{code_font_url}{Code Font URL. Defaults to https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700. Modifies the \code{@import url} elements.}

\item{code_font_family_fallback}{Code Font Fallback. Defaults to 'Lucida Console', Monaco. Modifies the \code{.remark-code, .remark-inline-code} classes.}

\item{extra_css}{A named list of CSS definitions each containing a named list
of CSS property-value pairs, i.e.
\code{list(".class-id" = list("css-property" = "value"))}}

\item{extra_fonts}{A list of additional fonts to import, each list element
can be either a URL as a character string or a call to
\code{\link{google_font}()}}

\item{outfile}{Customized xaringan CSS output file name, default is "xaringan-themer.css"}
}
\description{
The default xaringan theme with a single color used for color
accents on select elements (headers, bold text, etc.).
}
\seealso{
Other themes: \code{\link{duo_accent_inverse}},
  \code{\link{duo_accent}}, \code{\link{duo}},
  \code{\link{mono_accent_inverse}},
  \code{\link{mono_dark}}, \code{\link{mono_light}},
  \code{\link{solarized_dark}},
  \code{\link{solarized_light}},
  \code{\link{write_xaringan_theme}}

Other Monotone themes: \code{\link{mono_accent_inverse}},
  \code{\link{mono_dark}}, \code{\link{mono_light}}
}
\concept{Monotone themes}
\concept{themes}
