% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mono_accent_inverse.R
\name{mono_accent_inverse}
\alias{mono_accent_inverse}
\title{Monotone Accent Inverse Theme}
\usage{
mono_accent_inverse(base_color = "#3C989E", white_color = "#FFFFFF",
  black_color = darken_color(base_color, 0.9), text_color = white_color,
  header_color = base_color, background_color = black_color,
  link_color = base_color, text_bold_color = base_color,
  text_slide_number_color = base_color,
  code_highlight_color = "rgba(255,255,0,0.5)",
  code_inline_color = base_color, code_inline_background_color = NA,
  inverse_background_color = base_color, inverse_text_color = black_color,
  inverse_text_shadow = FALSE, inverse_header_color = black_color,
  title_slide_text_color = inverse_text_color,
  title_slide_background_color = inverse_background_color,
  title_slide_background_image = NA,
  left_column_subtle_color = apply_alpha(base_color, 0.6),
  left_column_selected_color = base_color,
  blockquote_left_color = apply_alpha(base_color, 0.5),
  table_border_color = "#666", table_row_border_color = "#ddd",
  table_row_even_background_color = darken_color(base_color, 0.7),
  text_size = "20px", header_h1_font_size = "55px",
  header_h2_font_size = "45px", header_h3_font_size = "35px",
  text_slide_number_font_size = "20px", code_inline_font_size = "1em",
  code_font_size = "0.9em", text_font_google = NULL,
  text_font_family = "'Droid Serif'", text_font_weight = "normal",
  text_font_url = "https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic",
  text_font_family_fallback = "'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'",
  text_font_base = "serif", header_font_google = NULL,
  header_font_family = "'Yanone Kaffeesatz'", header_font_weight = "normal",
  header_font_url = "https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz",
  code_font_google = NULL, code_font_family = "'Source Code Pro'",
  code_font_url = "https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700",
  code_font_family_fallback = "'Lucida Console', Monaco", extra_css = NULL,
  outfile = "xaringan-themer.css")
}
\arguments{
\item{base_color}{Monotone Base Color, works best with a light color., defaults to #3C989E}

\item{white_color}{Brightest color used, default is a very light version of \code{base_color}, defaults to #FFFFFF}

\item{black_color}{Darkest color used, default is a very dark, version of \code{base_color}, defaults to \code{darken_color(base_color, 0.9)}}

\item{text_color}{Text Color, defaults to \code{white_color}}

\item{header_color}{Header Color, defaults to \code{base_color}}

\item{background_color}{Slide Background Color, defaults to \code{black_color}}

\item{link_color}{Link Color, defaults to \code{base_color}}

\item{text_bold_color}{Bold Text Color, defaults to \code{base_color}}

\item{text_slide_number_color}{Slide Number Color, defaults to \code{base_color}}

\item{code_highlight_color}{Code Line Highlight, defaults to rgba(255,255,0,0.5)}

\item{code_inline_color}{Inline Code Color, defaults to \code{base_color}}

\item{code_inline_background_color}{Inline Code Background Color, defaults to NA}

\item{inverse_background_color}{Inverse Background Color, defaults to \code{base_color}}

\item{inverse_text_color}{Inverse Text Color, defaults to \code{black_color}}

\item{inverse_text_shadow}{Enables Shadow on text of inverse slides, defaults to \code{FALSE}}

\item{inverse_header_color}{Inverse Header Color, defaults to \code{black_color}}

\item{title_slide_text_color}{Title Slide Text Color, defaults to \code{inverse_text_color}}

\item{title_slide_background_color}{Title Slide Background Color, defaults to \code{inverse_background_color}}

\item{title_slide_background_image}{Title Slide Background Image URL, defaults to NA}

\item{left_column_subtle_color}{Left Column Text (not last), defaults to \code{apply_alpha(base_color, 0.6)}}

\item{left_column_selected_color}{Left Column Current Selection, defaults to \code{base_color}}

\item{blockquote_left_color}{Blockquote Left Border Color, defaults to \code{apply_alpha(base_color, 0.5)}}

\item{table_border_color}{Table top/bottom border, defaults to #666}

\item{table_row_border_color}{Table row inner bottom border, defaults to #ddd}

\item{table_row_even_background_color}{Table Even Row Background Color, defaults to \code{darken_color(base_color, 0.7)}}

\item{text_size}{Slide Body Text Size, defaults to 20px}

\item{header_h1_font_size}{h1 Header Text Size, defaults to 55px}

\item{header_h2_font_size}{h2 Header Text Size, defaults to 45px}

\item{header_h3_font_size}{h3 Header Text Size, defaults to 35px}

\item{text_slide_number_font_size}{Slide Number Text Size, defaults to 20px}

\item{code_inline_font_size}{Inline Code Text Size, defaults to 1em}

\item{code_font_size}{Code Text Size, defaults to 0.9em}

\item{text_font_google}{Use \code{google_font()} to specify body font, defaults to \code{NULL}}

\item{text_font_family}{Body Text Font Family, defaults to 'Droid Serif'}

\item{text_font_weight}{Body Text Font Weight, defaults to normal}

\item{text_font_url}{Body Text Font URL(s), defaults to https://fonts.googleapis.com/css?family=Droid+Serif:400,700,400italic}

\item{text_font_family_fallback}{Body Text Font Fallbacks, defaults to 'Palatino Linotype', 'Book Antiqua', Palatino, 'Microsoft YaHei', 'Songti SC'}

\item{text_font_base}{Body Text Base Font (Total Failure Fallback), defaults to serif}

\item{header_font_google}{Use \code{google_font()} to specify header font, defaults to \code{NULL}}

\item{header_font_family}{Header Font Family, defaults to 'Yanone Kaffeesatz'}

\item{header_font_weight}{Header Font Weight, defaults to normal}

\item{header_font_url}{Header Font URL, defaults to https://fonts.googleapis.com/css?family=Yanone+Kaffeesatz}

\item{code_font_google}{Use \code{google_font()} to specify code font, defaults to \code{NULL}}

\item{code_font_family}{Code Font Family, defaults to 'Source Code Pro'}

\item{code_font_url}{Code Font URL, defaults to https://fonts.googleapis.com/css?family=Source+Code+Pro:400,700}

\item{code_font_family_fallback}{Code Font Fallback, defaults to 'Lucida Console', Monaco}

\item{extra_css}{A named list of CSS definitions each containing a named list
of CSS property-value pairs, i.e.
\code{list(".class-id" = list("css-property" = "value"))}}

\item{outfile}{Customized xaringan CSS output file name, default is "xaringan-themer.css"}
}
\description{
An "inverted" default xaringan theme with a single color used
  for color accents on select elements (headers, bold text, etc.).
}
\seealso{
Other Monotone themes: \code{\link{mono_accent}},
  \code{\link{mono_dark}}, \code{\link{mono_light}}
}
