% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringanthemer-package.R
\docType{package}
\name{xaringanthemer-package}
\alias{xaringanthemer}
\alias{xaringanthemer-package}
\title{xaringanthemer: Custom 'Xaringan' CSS Themes}
\description{
Create beautifully color-cordinated and customized
    themes for your 'xaringan' slides, without writing any CSS. Complete
    your slide theme with 'ggplot2' themes that match the font and colors
    used in your slides.  Customized styles can be created directly in
    your slides' 'R Markdown' source file or in a separate external
    script.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gadenbuie/xaringanthemer}
  \item Report bugs at \url{https://github.com/gadenbuie/xaringanthemer/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{garrick@adenbuie.com} (\href{https://orcid.org/0000-0002-7111-0077}{ORCID})

}
\keyword{internal}
